# openapi.yaml
openapi: 3.1.0
info:
  title: Blue Economy Maritime Intelligence API (Deep Blue Project ISR Analytics)
  version: 1.1.0
  description: >
    Signed feed admission, incident lifecycle and Workstream F ISR analytics:
    multi-modal detection ingestion (AIS/SAR/RF/acoustic/optical) with
    mandatory classification labelling, clearance-scoped track/anomaly reads,
    dual-control outcome ledger and platform envelope v1.0 outbox topics
    maritime.isr.v1 / maritime.behaviour.v1 / maritime.outcome.v1.
    Authentication is loopback_trusted_proxy (local edge headers) or
    keycloak_rs256 (Bearer RS256 JWT verified against the realm JWKS).
    Authorization is fail-closed: every mutating route requires a designated
    verified-token role (isr-admin for feed-source administration,
    isr-feed-ingest for signed feed/detection admission, isr-analyst or
    isr-watch-officer for incident lifecycle, isr-analyst for outcome
    proposals, isr-adjudicator for outcome confirmations). Unrecognized,
    absent or read-only roles are denied with 403 on every mutation.
paths:
...
  /v1/outcomes:
    post:
      operationId: proposeOutcome
      description: First control of the dual-control outcome ledger; proposed_by is bound to the authenticated subject.
      security: [{trustedProxy: []}, {bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OutcomeProposal'}}}}
      responses: {'201': {description: Proposal recorded.}, '409': {description: Conflicting proposal evidence.}, '422': {description: Invalid proposal.}, '503': {description: Outcome ledger not configured.}}
  /v1/outcomes/{entry_id}/confirm:
    post:
      operationId: confirmOutcome
      description: Second control. The confirmer must differ from the proposer (DB-enforced); confirmation posts the TigerBeetle transfer, writes the immutable ledger entry and emits maritime.outcome.v1.
      security: [{trustedProxy: []}, {bearerAuth: []}]
      parameters: [{name: entry_id, in: path, required: true, schema: {type: string, pattern: '^[A-Za-z0-9._:-]{1,128}$'}}]
      responses: {'200': {description: Confirmed immutable entry.}, '404': {description: Unknown proposal.}, '409': {description: Dual-control violation or already confirmed.}}
  /v1/outcomes/aggregates:
    get:
      operationId: outcomeAggregates
      description: Classification-free outcome aggregates (counts and totals only); the only ISR read path open to the insurer-aggregator role.
      security: [{trustedProxy: []}, {bearerAuth: []}]
      responses: {'200': {description: Outcome aggregates.}, '403': {description: Insufficient role.}}
components:
  securitySchemes:
    trustedProxy: {type: apiKey, in: header, name: X-Authenticated-Principal, description: Local edge mode; edge also asserts X-Trusted-Proxy, X-Authenticated-Roles and X-Authenticated-Clearance.}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: Keycloak RS256 token; roles from realm_access.roles, clearance from the clearance claim.}
...
  schemas:
    Classification:
      type: string
      enum: [UNCLASSIFIED, RESTRICTED, CONFIDENTIAL, SECRET]
...
    OutcomeProposal:
      type: object
      additionalProperties: false
      required: [entry_id, entry_kind, incident_ref, classification, metric, quantity, unit]
      properties:
        entry_id: {type: string, pattern: '^[A-Za-z0-9._:-]{1,128}$'}
        entry_kind: {type: string, enum: [incident-verified, premium-delta]}
        incident_ref: {type: string, pattern: '^[A-Za-z0-9._:-]{1,128}$'}
        classification: {$ref: '#/components/schemas/Classification'}
        metric: {type: string, enum: [incident-reduction-count, premium-delta-basis-points]}
        quantity: {type: integer, minimum: 1}
        unit: {type: string, enum: [incidents, basis-points]}